% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindRateData.R
\name{bindRateData}
\alias{bindRateData}
\title{bindRateData}
\usage{
bindRateData(
  data.dir = "rate_data",
  batchSelection = NULL,
  norm.meth = "none",
  out.name = "all_rate_data.txt",
  returnData = TRUE
)
}
\arguments{
\item{data.dir}{character-path of files}

\item{batchSelection}{character vector of file names to include}

\item{norm.meth}{character}

\item{out.name}{character}

\item{returnData}{Boolean}
}
\description{
collects data exported from wave
}
