% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcInt.R
\name{calcInt}
\alias{calcInt}
\title{calcInt}
\usage{
calcInt(
  x,
  varname = "ocr",
  groupingVar = "samp_int",
  intVar = "interval",
  intervals = c("Int1", "Int2", "Int3", "Int4"),
  measure = "ocr",
  calcFunctions = c("mean", "min", "max", "min")
)
}
\arguments{
\item{x}{data frame}

\item{varname}{character, new name for column}

\item{groupingVar}{character, variable to group by before summarizing}

\item{intVar}{character-column name for intervals}

\item{intervals}{character vector of intervals to include}

\item{measure}{character, which column to summarize}

\item{calcFunctions}{character vector of the summary functions corresponding to the intervals, e.g. c("mean","min", "max", "min")}
}
\description{
calculates the summarized data over the time points in each interval
}
