% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_column_interval.R
\name{create_column_interval}
\alias{create_column_interval}
\title{create_column_interval}
\usage{
create_column_interval(
  df,
  timeVar = "time",
  Interval.list = list(Int0 = c(1), Int1 = c(2:3), Int2 = c(4:6), Int3 = c(7:10), Int4
    = c(11:14))
)
}
\arguments{
\item{df}{data frame}

\item{timeVar}{character-column name for time points (not time exact)}

\item{Interval.list}{list of vectors specifying time points to include for each interval}
}
\description{
Adds column with intervals spanning multiple time points
}
