% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_column_treatment.R
\name{create_column_treatment}
\alias{create_column_treatment}
\title{create_column_treatment}
\usage{
create_column_treatment(
  df,
  Treatment.list = list(Int1 = "basal", Int2 = "oligomycin", Int3 = "FCCP", Int4 =
    "Rot+AA", Int0 = "Start"),
  IntVar = "interval"
)
}
\arguments{
\item{df}{data frame}

\item{Treatment.list}{list with names equal to intervals}

\item{IntVar}{character-column name for intervals}
}
\description{
Adds column with inhibitors
}
