% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_outlier.R
\name{find_outlier}
\alias{find_outlier}
\title{find_outlier}
\usage{
find_outlier(x, mult = 1.5)
}
\arguments{
\item{x}{numeric}

\item{mult}{numeric - how many IQRs}
}
\description{
excludes outliers by IQR
}
