% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarySEwithin.R
\name{normDataWithin}
\alias{normDataWithin}
\title{normDataWithin}
\usage{
normDataWithin(
  data = NULL,
  idvar,
  measurevar,
  betweenvars = NULL,
  na.rm = FALSE,
  .drop = TRUE
)
}
\arguments{
\item{data}{a data frame.}

\item{idvar}{the name of a column that identifies each subject (or matched subjects)}

\item{measurevar}{the name of a column that contains the variable to be summarized}

\item{betweenvars}{a vector containing names of columns that are between-subjects variables}

\item{na.rm}{a boolean that indicates whether to ignore NA's}

\item{.drop}{Boolean}
}
\description{
normDataWithin
}
\author{
http://www.cookbook-r.com/
}
