% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeahorseR-package.R
\docType{package}
\name{SeahorseR-package}
\alias{SeahorseR}
\alias{SeahorseR-package}
\title{SeahorseR: R-based analysis of respirometry (Seahorse) data}
\description{
Assists with import and analysis of data from Seahorse Extracellular Flux xf96.
}
\author{
\strong{Maintainer}: Elizabeth Videlock \email{videlock@ucla.edu} (\href{https://orcid.org/0000-0001-8460-6189}{ORCID})

}
\keyword{internal}
