% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarySEwithin.R
\name{summarySE}
\alias{summarySE}
\title{summarySE}
\usage{
summarySE(
  data = NULL,
  measurevar,
  groupvars = NULL,
  na.rm = FALSE,
  conf.interval = 0.95,
  .drop = TRUE
)
}
\arguments{
\item{data}{a data frame.}

\item{measurevar}{character}

\item{groupvars}{character}

\item{na.rm}{a boolean that indicates whether to ignore NA's}

\item{conf.interval}{the percent range of the confidence interval (default is 95\%)}

\item{.drop}{Boolean}
}
\description{
stats for time repeated measures
}
\author{
http://www.cookbook-r.com/
}
