% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addNormalizationData.R
\name{add.norm}
\alias{add.norm}
\title{add normalization data}
\usage{
add.norm(raw.dat, norm.dat.dir, norm.meth, batch.col)
}
\arguments{
\item{raw.dat}{dataframe}

\item{norm.dat.dir}{character-path of files}

\item{norm.meth}{character}

\item{batch.col}{character, matches file names}
}
\description{
adds a column with normalization factors.
}
