% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellsPerInt.R
\name{wellsPerInt}
\alias{wellsPerInt}
\title{wellsPerInt}
\usage{
wellsPerInt(
  dat,
  measure = "ocr",
  wellVar = "well",
  groupingVars = c("AgeGrp", "interval", "mID"),
  tableVars = c("AgeGrp", "GT"),
  summaryColName = "nWells",
  intVar = "interval"
)
}
\arguments{
\item{dat}{data frame}

\item{measure}{character, what data column should be not na, defaults to "ocr"}

\item{wellVar}{character, column name for well, defaults to "well"}

\item{groupingVars}{character vector of variables to group by, defaults to c("AgeGrp","interval","mID")}

\item{tableVars}{character vector of variables to show results by, defaults to c("AgeGrp","GT")}

\item{summaryColName}{name for the result column, defaults to "nWells"}

\item{intVar}{character, name of column with intervals, defaults to "interval}
}
\description{
table of well counts per sample for each interval
}
